requirements/base.txt
django==2.2.12
django-crispy-forms==1.9.1
django-environ==0.4.5
django-extensions==2.2.9
mysqlclient==1.4.6
pip==20.1.1
Pillow==7.1.2
psycopg2-binary
pytz
six==1.15.0
sorl-thumbnail==12.6.3
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.5
pandas==1.2.0
pycodestyle==2.6.0
pylint==2.6.0
python-dateutil==2.8.1
python-decouple==3.4
requests==2.25.1
SQLAlchemy==1.3.22
sqlparse==0.4.1
toml==0.10.2
Unipath==1.1
urllib3==1.26.2
whitenoise==5.2.0
wrapt==1.12.1
